Move code to existing credential file and add credential server location
