+++ src/config/config.server.json
@@ -1,4 +1,4 @@
 {
-    "issuerServerUrl": "http://localhost:5108",
+    "issuerServerUrl": "https://email-verifier-backend.identity.rifos.org",
     "dataVaultUrl": "https://data-vault.identity.rifos.org"
 }
